Flappy-box moves up and down. Changes direction when button pushed

+++ local_src/game-1.0/colors.h
@@ -1,7 +1,7 @@
-#define WHITE  0xFFFF
-#define BLACK  0x0000
-#define PINK   0xF81D
-#define GREEN  0x3FE0
-#define YELLOW 0xFFC0
-#define BLUE   0x001F
-#define GRAY   0x736D
+#define WHITE  (uint32_t)0xFFFF
+#define BLACK  (uint32_t)0x0000
+#define PINK   (uint32_t)0xF81D
+#define GREEN  (uint32_t)0x3FE0
+#define YELLOW (uint32_t)0xFFC0
+#define BLUE   (uint32_t)0x001F
+#define GRAY   (uint32_t)0x736D
